Package folder; package.json aligned; new install shell script.
