+++ ht16k33/__init__.py
@@ -1,0 +1,14 @@
+__author__ = 'Tony Smith'
+__credits__ = ['Tony Smith', 'Ubi de Feo', 'akbiocca']
+__license__ = "MIT"
+__version__ = "3.6.0"
+__maintainer__ = "Tony Smith"
+
+from .ht16k33 import *
+from .ht16k33matrix import *
+from .ht16k33matrixcolour import *
+from .ht16k33matrixfeatherwing import *
+from .ht16k33segment import *
+from .ht16k33segment14 import *
+from .ht16k33segmentbig import *
+from .ht16k33segmentgen import *
