+++ ht16k33/ht16k33segment.py
@@ -1,9 +1,9 @@
 # Import the base class
-from ht16k33 import HT16K33
+from .ht16k33 import HT16K33
 
-class HT16K33SegmentBig(HT16K33):
+class HT16K33Segment(HT16K33):
     """
-    Micro/Circuit Python class for the Adafruit 1.2-in 4-digit,
+    Micro/Circuit Python class for the Adafruit 0.56-in 4-digit,
     7-segment LED matrix backpack and equivalent Featherwing.
 
     Bus:        I2C
@@ -17,56 +17,51 @@ class HT16K33SegmentBig(HT16K33):
     HT16K33_SEGMENT_COLON_ROW = 0x04
     HT16K33_SEGMENT_MINUS_CHAR = 0x10
     HT16K33_SEGMENT_DEGREE_CHAR = 0x11
-    HT16K33_SEGMENT_SPACE_CHAR = 0x13
-
-    COLON_CENTRE = 0x02
-    COLON_LEFT_UPPER = 0x04
-    COLON_LEFT_LOWER = 0x08
-    COLON_LEFT = 0x0C
-    DECIMAL_POINT = 0x10
+    HT16K33_SEGMENT_SPACE_CHAR = 0x12
 
     # The positions of the segments within the buffer
     POS = (0, 2, 6, 8)
 
     # Bytearray of the key alphanumeric characters we can show:
-    # 0-9, A-F, minus, degree
+    # 0-9, A-F, minus, degree, space
     CHARSET = b'\x3F\x06\x5B\x4F\x66\x6D\x7D\x07\x7F\x6F\x5F\x7C\x58\x5E\x7B\x71\x40\x63\x00'
 
     # *********** CONSTRUCTOR **********
 
     def __init__(self, i2c, i2c_address=0x70):
         self.buffer = bytearray(16)
-        self.point_pattern = 0x00
-        super(HT16K33SegmentBig, self).__init__(i2c, i2c_address)
+        self.is_rotated = False
+        super(HT16K33Segment, self).__init__(i2c, i2c_address)
 
     # *********** PUBLIC METHODS **********
 
-    def set_colon(self, pattern=0x02):
+    def rotate(self):
         """
-        Set or unset the segment LED display's colon and decimal point lights.
-
-        This method updates the display buffer, but does not send the buffer to the display itself.
-        Call 'draw()' to render the buffer on the display.
-
-        Args:
-            patter (int) An integer indicating which elements to light (OR the values required). Default: 0x00
-                         0x00: no colon
-                         0x02: centre colon
-                         0x04: left colon, lower dot
-                         0x08: left colon, upper dot
-                         0x10: decimal point (upper)
+        Rotate/flip the segment display.
 
         Returns:
             The instance (self)
         """
-        # Bail on incorrect pattern values
-        assert 0 <= pattern < 0x1F, "ERROR - bad patter passed to set_colon()"
-
-        self.point_pattern = pattern
-        self.buffer[self.HT16K33_SEGMENT_COLON_ROW] = pattern
+        self.is_rotated = not self.is_rotated
         return self
 
-    def set_glyph(self, glyph, digit=0):
+    def set_colon(self, is_set=True):
+        """
+        Set or unset the display's central colon symbol.
+
+        This method updates the display buffer, but does not send the buffer to the display itself.
+        Call 'update()' to render the buffer on the display.
+
+        Args:
+            isSet (bool): Whether the colon is lit (True) or not (False). Default: True.
+
+        Returns:
+            The instance (self)
+        """
+        self.buffer[self.HT16K33_SEGMENT_COLON_ROW] = 0x02 if is_set is True else 0x00
+        return self
+
+    def set_glyph(self, glyph, digit=0, has_dot=False):
         """
         Present a user-defined character glyph at the specified digit.
 
@@ -89,18 +84,20 @@ class HT16K33SegmentBig(HT16K33):
         Args:
             glyph (int):   The glyph pattern.
             digit (int):   The digit to show the glyph. Default: 0 (leftmost digit).
+            has_dot (bool): Whether the decimal point to the right of the digit should be lit. Default: False.
 
         Returns:
             The instance (self)
         """
         # Bail on incorrect row numbers or character values
         assert 0 <= digit < 4, "ERROR - Invalid digit (0-3) set in set_glyph()"
-        assert 0 <= glyph < 0x80, "ERROR - Invalid glyph (0x00-0xFF) set in set_glyph()"
+        assert 0 <= glyph < 0x80, "ERROR - Invalid glyph (0x00-0x80) set in set_glyph()"
 
-        self.buffer[self.POS[digit]] = glyph & 0x7F
+        self.buffer[self.POS[digit]] = glyph
+        if has_dot is True: self.buffer[self.POS[digit]] |= 0x80
         return self
 
-    def set_number(self, number, digit=0):
+    def set_number(self, number, digit=0, has_dot=False):
         """
         Present single decimal value (0-9) at the specified digit.
 
@@ -110,6 +107,7 @@ class HT16K33SegmentBig(HT16K33):
         Args:
             number (int):  The number to show.
             digit (int):   The digit to show the number. Default: 0 (leftmost digit).
+            has_dot (bool): Whether the decimal point to the right of the digit should be lit. Default: False.
 
         Returns:
             The instance (self)
@@ -118,9 +116,9 @@ class HT16K33SegmentBig(HT16K33):
         assert 0 <= digit < 4, "ERROR - Invalid digit (0-3) set in set_number()"
         assert 0 <= number < 10, "ERROR - Invalid value (0-9) set in set_number()"
 
-        return self.set_character(str(number), digit)
+        return self.set_character(str(number), digit, has_dot)
 
-    def set_character(self, char, digit=0):
+    def set_character(self, char, digit=0, has_dot=False):
         """
         Present single alphanumeric character at the specified digit.
 
@@ -129,7 +127,7 @@ class HT16K33SegmentBig(HT16K33):
         Other characters can be defined and presented using 'set_glyph()'.
 
         This method updates the display buffer, but does not send the buffer to the display itself.
-        Call 'draw()' to render the buffer on the display.
+        Call 'update()' to render the buffer on the display.
 
         Args:
             char (string):  The character to show.
@@ -159,4 +157,31 @@ class HT16K33SegmentBig(HT16K33):
         assert char_val != 0xFF, "ERROR - Invalid char string set in set_character()"
 
         self.buffer[self.POS[digit]] = self.CHARSET[char_val]
+        if has_dot is True: self.buffer[self.POS[digit]] |= 0x80
         return self
+
+    def draw(self):
+        """
+        Writes the current display buffer to the display itself.
+
+        Call this method after updating the buffer to update
+        the LED itself. Rotation handled here.
+        """
+        if self.is_rotated:
+            # Swap digits 0,3 and 1,2
+            a = self.buffer[self.POS[0]]
+            self.buffer[self.POS[0]] = self.buffer[self.POS[3]]
+            self.buffer[self.POS[3]] = a
+
+            a = self.buffer[self.POS[1]]
+            self.buffer[self.POS[1]] = self.buffer[self.POS[2]]
+            self.buffer[self.POS[2]] = a
+
+            # Rotate each digit
+            for i in range(0, 4):
+                a = self.buffer[self.POS[i]]
+                b = (a & 0x07) << 3
+                c = (a & 0x38) >> 3
+                a &= 0xC0
+                self.buffer[self.POS[i]] = (a | b | c)
+        self._render()
