+++ ht16k33/ht16k33segmentbig.py
@@ -1,11 +1,10 @@
 # Import the base class
-from ht16k33 import HT16K33
+from .ht16k33 import HT16K33
 
-class HT16K33SegmentGen(HT16K33):
+class HT16K33SegmentBig(HT16K33):
     """
-    Micro/Circuit Python class for a generic 1-8-digit, 7-segment display.
-    It assumes each digit has a decimal point, but there are no other
-    symbol LEDs included.
+    Micro/Circuit Python class for the Adafruit 1.2-in 4-digit,
+    7-segment LED matrix backpack and equivalent Featherwing.
 
     Bus:        I2C
     Author:     Tony Smith (@smittytone)
@@ -15,37 +14,59 @@ class HT16K33SegmentGen(HT16K33):
 
     # *********** CONSTANTS **********
 
+    HT16K33_SEGMENT_COLON_ROW = 0x04
     HT16K33_SEGMENT_MINUS_CHAR = 0x10
     HT16K33_SEGMENT_DEGREE_CHAR = 0x11
-    HT16K33_SEGMENT_SPACE_CHAR = 0x12
+    HT16K33_SEGMENT_SPACE_CHAR = 0x13
+
+    COLON_CENTRE = 0x02
+    COLON_LEFT_UPPER = 0x04
+    COLON_LEFT_LOWER = 0x08
+    COLON_LEFT = 0x0C
+    DECIMAL_POINT = 0x10
+
+    # The positions of the segments within the buffer
+    POS = (0, 2, 6, 8)
 
     # Bytearray of the key alphanumeric characters we can show:
-    # 0-9, A-F, minus, degree, space
+    # 0-9, A-F, minus, degree
     CHARSET = b'\x3F\x06\x5B\x4F\x66\x6D\x7D\x07\x7F\x6F\x5F\x7C\x58\x5E\x7B\x71\x40\x63\x00'
 
     # *********** CONSTRUCTOR **********
 
-    def __init__(self, i2c, i2c_address=0x70, digits=8):
+    def __init__(self, i2c, i2c_address=0x70):
         self.buffer = bytearray(16)
-        self.is_rotated = False
-        # Check digits specified (must be 1 - 8)
-        assert 0 < digits < 9, "ERROR - Invalid number of digits (1-8) in HT16K33Segment8()"
-        self.max_digits = digits
-        super(HT16K33SegmentGen, self).__init__(i2c, i2c_address)
+        self.point_pattern = 0x00
+        super(HT16K33SegmentBig, self).__init__(i2c, i2c_address)
 
     # *********** PUBLIC METHODS **********
 
-    def rotate(self):
+    def set_colon(self, pattern=0x02):
         """
-        Rotate/flip the segment display.
+        Set or unset the segment LED display's colon and decimal point lights.
+
+        This method updates the display buffer, but does not send the buffer to the display itself.
+        Call 'draw()' to render the buffer on the display.
+
+        Args:
+            patter (int) An integer indicating which elements to light (OR the values required). Default: 0x00
+                         0x00: no colon
+                         0x02: centre colon
+                         0x04: left colon, lower dot
+                         0x08: left colon, upper dot
+                         0x10: decimal point (upper)
 
         Returns:
             The instance (self)
         """
-        self.is_rotated = not self.is_rotated
+        # Bail on incorrect pattern values
+        assert 0 <= pattern < 0x1F, "ERROR - bad patter passed to set_colon()"
+
+        self.point_pattern = pattern
+        self.buffer[self.HT16K33_SEGMENT_COLON_ROW] = pattern
         return self
 
-    def set_glyph(self, glyph, digit=0, has_dot=False):
+    def set_glyph(self, glyph, digit=0):
         """
         Present a user-defined character glyph at the specified digit.
 
@@ -68,20 +89,18 @@ class HT16K33SegmentGen(HT16K33):
         Args:
             glyph (int):   The glyph pattern.
             digit (int):   The digit to show the glyph. Default: 0 (leftmost digit).
-            has_dot (bool): Whether the decimal point to the right of the digit should be lit. Default: False.
 
         Returns:
             The instance (self)
         """
         # Bail on incorrect row numbers or character values
-        assert 0 <= digit < self.max_digits, "ERROR - Invalid digit set in set_glyph()"
-        assert 0 <= glyph < 0x80, "ERROR - Invalid glyph (0x00-0x80) set in set_glyph()"
+        assert 0 <= digit < 4, "ERROR - Invalid digit (0-3) set in set_glyph()"
+        assert 0 <= glyph < 0x80, "ERROR - Invalid glyph (0x00-0xFF) set in set_glyph()"
 
-        self.buffer[digit << 1] = glyph
-        if has_dot is True: self.buffer[digit << 1] |= 0x80
+        self.buffer[self.POS[digit]] = glyph & 0x7F
         return self
 
-    def set_number(self, number, digit=0, has_dot=False):
+    def set_number(self, number, digit=0):
         """
         Present single decimal value (0-9) at the specified digit.
 
@@ -91,18 +110,17 @@ class HT16K33SegmentGen(HT16K33):
         Args:
             number (int):  The number to show.
             digit (int):   The digit to show the number. Default: 0 (leftmost digit).
-            has_dot (bool): Whether the decimal point to the right of the digit should be lit. Default: False.
 
         Returns:
             The instance (self)
         """
         # Bail on incorrect row numbers or character values
-        assert 0 <= digit < self.max_digits, "ERROR - Invalid digit set in set_number()"
+        assert 0 <= digit < 4, "ERROR - Invalid digit (0-3) set in set_number()"
         assert 0 <= number < 10, "ERROR - Invalid value (0-9) set in set_number()"
 
-        return self.set_character(str(number), digit, has_dot)
+        return self.set_character(str(number), digit)
 
-    def set_character(self, char, digit=0, has_dot=False):
+    def set_character(self, char, digit=0):
         """
         Present single alphanumeric character at the specified digit.
 
@@ -111,7 +129,7 @@ class HT16K33SegmentGen(HT16K33):
         Other characters can be defined and presented using 'set_glyph()'.
 
         This method updates the display buffer, but does not send the buffer to the display itself.
-        Call 'update()' to render the buffer on the display.
+        Call 'draw()' to render the buffer on the display.
 
         Args:
             char (string):  The character to show.
@@ -122,7 +140,7 @@ class HT16K33SegmentGen(HT16K33):
             The instance (self)
         """
         # Bail on incorrect row numbers
-        assert 0 <= digit < self.max_digits, "ERROR - Invalid digit set in set_character()"
+        assert 0 <= digit < 4, "ERROR - Invalid digit set in set_character()"
 
         char = char.lower()
         char_val = 0xFF
@@ -140,43 +158,5 @@ class HT16K33SegmentGen(HT16K33):
         # Bail on incorrect character values
         assert char_val != 0xFF, "ERROR - Invalid char string set in set_character()"
 
-        self.buffer[digit << 1] = self.CHARSET[char_val]
-        if has_dot is True: self.buffer[digit << 1] |= 0x80
+        self.buffer[self.POS[digit]] = self.CHARSET[char_val]
         return self
-
-    def draw(self):
-        """
-        Writes the current display buffer to the display itself.
-
-        Call this method after updating the buffer to update
-        the LED itself. Rotation handled here.
-        """
-        if self.is_rotated:
-            # Preserve the unrotated buffer
-            tmpbuffer = bytearray(16)
-            for i in range(0, self.max_digits << 1):
-                tmpbuffer[i] = self.buffer[i]
-            # Swap digits 0,(max - 1), 1,(max - 2) etc
-            if self.max_digits > 1:
-                for i in range(0, (self.max_digits >> 1)):
-                    right = (self.max_digits - i - 1) << 1
-                    left = i << 1
-                    if left != right:
-                        a = self.buffer[left]
-                        self.buffer[left] = self.buffer[right]
-                        self.buffer[right] = a
-
-            # Flip each digit
-            for i in range(0, self.max_digits):
-                a = self.buffer[i << 1]
-                b = (a & 0x07) << 3
-                c = (a & 0x38) >> 3
-                a &= 0xC0
-                self.buffer[i << 1] = (a | b | c)
-            self._render()
-            # Restore the buffer
-            for i in range(0, self.max_digits << 1):
-                self.buffer[i] = tmpbuffer[i]
-        else:
-            self._render()
-
